started my courses

+++ src/Glava_1/Task_1_8/Ages.java
@@ -6,7 +6,7 @@ public class Ages {
         byte ag2 = 10;
         System.out.println(getAgeDiff(ag1, ag2));
     }
-    public static int  getAgeDiff(byte age1, byte age2) {
+    public static byte  getAgeDiff(byte age1, byte age2) {
         return (byte) Math.abs(age2-age1);
     }
 }
